20220804 inp

+++ zantt-api-backend-net/Zantt/Services/ProjectService.cs
@@ -1,4 +1,5 @@
-﻿using Zantt.Entities;
+﻿using System.Diagnostics;
+using Zantt.Entities;
 using Zantt.Exceptions;
 using Zantt.Repositories;
 
@@ -24,11 +25,21 @@ public class ProjectService
 
     public ProjectEntity? GetProject(string projectId)
     {
+        if (projectId == null)
+        {
+            throw new WellKnownApiException($"{nameof(projectId)} is null", "INVALID_PARAM");
+        }
+
         return projectRepository.GetProjectByProjectId(projectId);
     }
 
     public ProjectEntity? AddProject(string name)
     {
+        if (name == null)
+        {
+            throw new WellKnownApiException($"{nameof(name)} is null", "INVALID_PARAM");
+        }
+
         var projectId = Guid.NewGuid().ToString();
 
         var project = new ProjectEntity()
@@ -42,6 +53,15 @@ public class ProjectService
 
     public ProjectEntity? UpdateProject(string projectId, string name)
     {
+        if (projectId == null)
+        {
+            throw new WellKnownApiException($"{nameof(projectId)} is null", "INVALID_PARAM");
+        }
+        if (name == null)
+        {
+            throw new WellKnownApiException($"{nameof(name)} is null", "INVALID_PARAM");
+        }
+
         var project = projectRepository.UpdateProjectByProjectId(projectId, name);
         if (project == null)
         {
@@ -53,6 +73,11 @@ public class ProjectService
 
     public void DeleteProject(string projectId)
     {
+        if (projectId == null)
+        {
+            throw new WellKnownApiException($"{nameof(projectId)} is null", "INVALID_PARAM");
+        }
+
         projectRepository.DeleteProjectByProjectId(projectId);
     }
 }
